feat: work for liulab

+++ FreeSurfer/scripts/RescaleVolumeAndSurface.sh
@@ -1,16 +1,17 @@
 #! /bin/bash
 set -e
 
-Usage_exit (){
-echo "RescaleVolumeAndSurface.sh <SubjectDIR> <SubjectID> <RescaleVolumeTransform> <T1wImage>"
-exit 0;
+Usage_exit() {
+  echo "RescaleVolumeAndSurface.sh <SubjectDIR> <SubjectID> <RescaleVolumeTransform> <T1wImage>"
+  exit 0
 }
-if [ "$4" = "" ] ; then Usage_exit ;fi
+if [ "$4" = "" ]; then Usage_exit; fi
 
 SubjectDIR=$1
 SubjectID=$2
 RescaleVolumeTransform=$3
 T1wImage=$4
+NoT2wData=$5
 
 mri_convert -rt cubic -at "$RescaleVolumeTransform".xfm -rl "$T1wImage" "$SubjectDIR"/"$SubjectID"_1mm/mri/rawavg.mgz "$SubjectDIR"/"$SubjectID"/mri/rawavg.mgz
 mri_convert "$SubjectDIR"/"$SubjectID"/mri/rawavg.mgz "$SubjectDIR"/"$SubjectID"/mri/rawavg.nii.gz
@@ -32,20 +33,29 @@ mri_convert "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.nii.gz "$SubjectDIR"
 mri_convert "$SubjectDIR"/"$SubjectID"/mri/orig.nii.gz "$SubjectDIR"/"$SubjectID"/mri/orig.mgz
 
 # for hemisphere
-mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz --tval white_temp --hemi lh
+# NOTE: 可能是版本问题，`--tval-xyz`后面必须跟一个volume文件，因此修改。
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval white_temp --hemi lh
 mv "$SubjectDIR"/"$SubjectID"_1mm/surf/lh.white_temp "$SubjectDIR"/"$SubjectID"/surf/lh.white
-mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz --tval white_temp --hemi rh
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval white_temp --hemi rh
 mv "$SubjectDIR"/"$SubjectID"_1mm/surf/rh.white_temp "$SubjectDIR"/"$SubjectID"/surf/rh.white
-mri_surf2surf --s "$SubjectID"_1mm --sval-xyz pial --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz --tval pial_temp --hemi lh
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz pial --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval pial_temp --hemi lh
 mv "$SubjectDIR"/"$SubjectID"_1mm/surf/lh.pial_temp "$SubjectDIR"/"$SubjectID"/surf/lh.pial
-mri_surf2surf --s "$SubjectID"_1mm --sval-xyz pial --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz --tval pial_temp --hemi rh
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz pial --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval pial_temp --hemi rh
 mv "$SubjectDIR"/"$SubjectID"_1mm/surf/rh.pial_temp "$SubjectDIR"/"$SubjectID"/surf/rh.pial
 
-mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white.deformed --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz --tval white.deformed_temp --hemi lh
+# NOTE: 可能是版本问题，`--tval-xyz`后面必须跟一个volume文件，因此修改。
+# 在${subj}文件夹中创建左右脑的?h.white.deformed文件（这个文件和T1w_acpc_dc_restore_brain.nii.gz是一样分辨率的。即，不是Freesurfer中的1x1x1）
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white.deformed --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval white.deformed_temp --hemi lh
 mv "$SubjectDIR"/"$SubjectID"_1mm/surf/lh.white.deformed_temp "$SubjectDIR"/"$SubjectID"/surf/lh.white.deformed
-mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white.deformed --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz --tval white.deformed_temp --hemi rh
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz white.deformed --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval white.deformed_temp --hemi rh
 mv "$SubjectDIR"/"$SubjectID"_1mm/surf/rh.white.deformed_temp "$SubjectDIR"/"$SubjectID"/surf/rh.white.deformed
 
+# NOTE: 可能是版本问题，`--tval-xyz`后面必须跟一个volume文件，因此修改。
+# 在${subj}文件夹中创建左右脑的?h.pial.deformed文件（这个文件和T1w_acpc_dc_restore_brain.nii.gz是一样分辨率的。即，不是Freesurfer中的1x1x1）
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz pial.deformed --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval pial.deformed_temp --hemi lh # 在${subj}_1mm/surf文件夹中创建一个lh.pial.deformed_temp文件，这个lh.pial.deformed_temp文件已经是T1w_acpc_dc_restore_brain.nii.gz分辨率的
+mv "$SubjectDIR"/"$SubjectID"_1mm/surf/lh.pial.deformed_temp "$SubjectDIR"/"$SubjectID"/surf/lh.pial.deformed                                                                                                                                              # 把{subj}_1mm/surf文件夹中的lh.pial.deformed_temp文件移动到{subj}/surf文件夹中
+mri_surf2surf --s "$SubjectID"_1mm --sval-xyz pial.deformed --reg-inv "$RescaleVolumeTransform".dat "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval-xyz "$SubjectDIR"/"$SubjectID"/mri/brain.finalsurfs.mgz --tval pial.deformed_temp --hemi rh
+mv "$SubjectDIR"/"$SubjectID"_1mm/surf/rh.pial.deformed_temp "$SubjectDIR"/"$SubjectID"/surf/rh.pial.deformed
 
 cp "$SubjectDIR"/"$SubjectID"_1mm/surf/lh.sphere "$SubjectDIR"/"$SubjectID"/surf/lh.sphere
 cp "$SubjectDIR"/"$SubjectID"_1mm/surf/rh.sphere "$SubjectDIR"/"$SubjectID"/surf/rh.sphere
@@ -62,18 +72,28 @@ cp "$SubjectDIR"/"$SubjectID"_1mm/label/lh.aparc.annot "$SubjectDIR"/"$SubjectID
 cp "$SubjectDIR"/"$SubjectID"_1mm/label/rh.aparc.annot "$SubjectDIR"/"$SubjectID"/label/rh.aparc.annot
 
 # For Volume and matrices
-cp "$RescaleVolumeTransform".mat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/fs2real.mat
-convert_xfm -omat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/real2fs.mat -inverse "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/fs2real.mat
-convert_xfm -omat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/temp.mat -concat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/T2wtoT1w.mat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/real2fs.mat
-convert_xfm -omat "$SubjectDIR"/"$SubjectID"/mri/transforms/T2wtoT1w.mat -concat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/fs2real.mat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/temp.mat
-rm "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/temp.mat
+cp "$RescaleVolumeTransform".mat \
+  "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/fs2real.mat
+convert_xfm -omat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/real2fs.mat \
+  -inverse "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/fs2real.mat
+if [[ $NoT2wData -ne 1 ]]; then
+  convert_xfm -omat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/temp.mat \
+    -concat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/T2wtoT1w.mat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/real2fs.mat
+  convert_xfm -omat "$SubjectDIR"/"$SubjectID"/mri/transforms/T2wtoT1w.mat \
+    -concat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/fs2real.mat "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/temp.mat
+  rm "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/temp.mat
+else
+  echo " "
+  echo "没有T2w数据，以下步骤将跳过：在FS/<subj>/mri/transforms下生成T2wtoT1w.mat！"
+fi
+
 cp "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/eye.dat "$SubjectDIR"/"$SubjectID"/mri/transforms/eye.dat
-if [ `head -n 1 "$SubjectDIR"/"$SubjectID"/mri/transforms/eye.dat | tail -c 5` != "_1mm" ] ; then # avoid adding too many "_1mm" - Takuya Hayashi
-	cat "$SubjectDIR"/"$SubjectID"/mri/transforms/eye.dat | sed "s/${SubjectID}/${SubjectID}_1mm/g" > "$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/eye.dat
+if [ $(head -n 1 "$SubjectDIR"/"$SubjectID"/mri/transforms/eye.dat | tail -c 5) != "_1mm" ]; then # avoid adding too many "_1mm" - Takuya Hayashi
+  cat "$SubjectDIR"/"$SubjectID"/mri/transforms/eye.dat | sed "s/${SubjectID}/${SubjectID}_1mm/g" >"$SubjectDIR"/"$SubjectID"_1mm/mri/transforms/eye.dat
 fi
 
 # For hemisphere
-for hemisphere in l r ; do
+for hemisphere in l r; do
   cp "$SubjectDIR"/"$SubjectID"_1mm/surf/${hemisphere}h.thickness "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.thickness
   #mris_convert "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.white "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.white.surf.gii
   #mris_convert "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.pial "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.pial.surf.gii
@@ -83,11 +103,26 @@ for hemisphere in l r ; do
   #mris_convert -c "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.thickness.shape.gii "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.white "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.thickness.asc
   #rm "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.white.surf.gii "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.pial.surf.gii "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.thickness.shape.gii "$SubjectDIR"/"$SubjectID"/surf/${hemisphere}h.roi.shape.gii
 done
+
 surf="${SubjectDIR}/${SubjectID}/surf"
+
+# hemi="lh"
+# matlab -nojvm -nosplash <<M_PROG
+# addpath $HCPPIPEDIR/global/matlab
+# addpath('/usr/local/NHPHCP/matlab')
+# corticalthickness('${surf}','${hemi}');
+# M_PROG
+# hemi="rh"
+# matlab -nojvm -nosplash <<M_PROG
+# addpath $HCPPIPEDIR/global/matlab
+# addpath('/usr/local/NHPHCP/matlab')
+# corticalthickness('${surf}','${hemi}');
+# M_PROG
+
+# NOTE: 需要将$FREESURFER_HOME/matlab添加到MATLAB路径中
 hemi="lh"
 matlab -nojvm -nosplash <<M_PROG
 addpath $HCPPIPEDIR/global/matlab
-addpath('/usr/local/NHPHCP/matlab')
 corticalthickness('${surf}','${hemi}');
 M_PROG
 hemi="rh"
